+++ src/renderer/theme/themes/solarized-light.ts
@@ -66,6 +66,7 @@ export const solarizedLight: ThemeDefinition = {
         "--color-misc-green": "#859900",
         "--color-misc-red": "#dc322f",
         "--color-misc-yellow": "#b58900",
+        "--color-misc-orange": "#cb4b16",
         "--color-misc-vlc": "#ff8800",
 
         // error